requirements.txt
pyserial>=3.4
numpy>=1.15.4
matplotlib>=2.2.3
